+++ uvicorn/config.py
@@ -87,9 +87,7 @@ LOOP_SETUPS: Dict[LoopSetupType, Optional[str]] = {
 }
 INTERFACES: List[InterfaceType] = ["auto", "asgi3", "asgi2", "wsgi"]
 
-
 SSL_PROTOCOL_VERSION: int = ssl.PROTOCOL_TLS_SERVER
-
 
 LOGGING_CONFIG: Dict[str, Any] = {
     "version": 1,
@@ -118,7 +116,7 @@ LOGGING_CONFIG: Dict[str, Any] = {
         },
     },
     "loggers": {
-        "uvicorn": {"handlers": ["default"], "level": "INFO"},
+        "uvicorn": {"handlers": ["default"], "level": "INFO", "propagate": False},
         "uvicorn.error": {"level": "INFO"},
         "uvicorn.access": {"handlers": ["access"], "level": "INFO", "propagate": False},
     },
@@ -163,7 +161,6 @@ def is_dir(path: Path) -> bool:
 def resolve_reload_patterns(
     patterns_list: List[str], directories_list: List[str]
 ) -> Tuple[List[str], List[Path]]:
-
     directories: List[Path] = list(set(map(Path, directories_list.copy())))
     patterns: List[str] = patterns_list.copy()
 
@@ -380,6 +377,9 @@ class Config:
             )
         else:
             self.forwarded_allow_ips = forwarded_allow_ips
+
+        if self.reload and self.workers > 1:
+            logger.warning('"workers" flag is ignored when reloading is enabled.')
 
     @property
     def asgi_version(self) -> Literal["2.0", "3.0"]: