+++ uvicorn/main.py
@@ -578,7 +578,7 @@ def run(
         Multiprocess(config, target=server.run, sockets=[sock]).run()
     else:
         server.run()
-    if config.uds:
+    if config.uds and os.path.exists(config.uds):
         os.remove(config.uds)  # pragma: py-win32
 
     if not server.started and not config.should_reload and config.workers == 1: